+++ locals-naming.tf
@@ -1,22 +1,9 @@
 locals {
   name_prefix = lower(var.name_prefix)
   name_suffix = lower(var.name_suffix)
-  caf_naming_resources = [
-    "azurerm_virtual_wan",
-    "azurerm_express_route_gateway",
-    "azurerm_express_route_circuit",
-    "azurerm_firewall",
-    "azurerm_public_ip",
-    "azurerm_firewall_ip_configuration"
-  ]
 
-  vwan_name = coalesce(var.custom_vwan_name, azurecaf_name.caf["azurerm_virtual_wan"].result)
-  ergw_name = coalesce(var.custom_express_route_gateway_name, azurecaf_name.caf["azurerm_express_route_gateway"].result)
-  erc_name  = coalesce(var.custom_express_route_circuit_name, azurecaf_name.caf["azurerm_express_route_circuit"].result)
-  fw_name   = coalesce(var.custom_firewall_name, azurecaf_name.caf["azurerm_firewall"].result)
-
+  vwan_name         = coalesce(var.custom_vwan_name, azurecaf_name.virtual_wan_caf.result)
   default_naming    = "${local.name_prefix}${var.stack}-${var.client_name}-${var.location_short}-${var.environment}"
   default_vhub_name = "${local.default_naming}-vhub"
 
-  vhub_name = coalesce(var.custom_virtual_hub_name, local.default_vhub_name)
 }