+++ modules/express-route/variables.tf
@@ -1,0 +1,146 @@
+variable "client_name" {
+  description = "Name of client."
+  type        = string
+}
+
+variable "environment" {
+  description = "Name of application's environment."
+  type        = string
+}
+
+variable "stack" {
+  description = "Name of application's stack."
+  type        = string
+}
+
+variable "resource_group_name" {
+  description = "Name of the application's resource group."
+  type        = string
+}
+
+variable "location" {
+  description = "Azure location."
+  type        = string
+}
+
+variable "location_short" {
+  description = "Short string for Azure location."
+  type        = string
+}
+
+variable "name_prefix" {
+  description = "Prefix for generated resources names."
+  type        = string
+  default     = ""
+}
+
+variable "name_suffix" {
+  description = "Suffix for the generated resources names."
+  type        = string
+  default     = ""
+}
+
+variable "name_slug" {
+  description = "Slug to use with the generated resources names."
+  type        = string
+  default     = ""
+}
+
+variable "logs_destinations_ids" {
+  description = "List of destination resources IDs for logs diagnostic destination. Can be Storage Account, Log Analytics Workspace and Event Hub. No more than one of each can be set."
+  type        = list(string)
+}
+
+variable "custom_express_route_gateway_name" {
+  description = "Custom Express Route Gateway name"
+  type        = string
+  default     = null
+}
+
+variable "express_route_gateway_tags" {
+  description = "Extra tags for Express Route Gateway"
+  type        = map(string)
+  default     = {}
+}
+
+variable "custom_express_route_circuit_name" {
+  description = "Custom Express Route Circuit name"
+  type        = string
+  default     = null
+}
+
+variable "express_route_gateway_scale_unit" {
+  description = "The number of scale unit with which to provision the Express Route Gateway."
+  type        = number
+  default     = 1
+}
+variable "express_route_circuit_peering_location" {
+  description = "Express Route Circuit peering location."
+  type        = string
+  default     = null
+}
+
+variable "express_route_circuit_bandwidth_in_mbps" {
+  description = "The bandwith in Mbps of the Express Route Circuit being created on the Service Provider"
+  type        = number
+  default     = null
+}
+
+variable "express_route_circuit_service_provider" {
+  description = "The name of the Express Route Circuit Service Provider."
+  type        = string
+  default     = null
+}
+
+variable "express_route_sku" {
+  description = "Express Route SKU"
+  type = object({
+    tier   = string,
+    family = string
+  })
+  default = {
+    tier   = "Premium"
+    family = "MeteredData"
+  }
+}
+
+variable "express_route_private_peering_enabled" {
+  description = "Enable Express Route Circuit Private Peering"
+  type        = bool
+  default     = false
+}
+
+variable "express_route_circuit_private_peering_primary_peer_address_prefix" {
+  description = "Primary peer address prefix for Express Route Circuit private peering"
+  type        = string
+  default     = null
+}
+
+variable "express_route_circuit_private_peering_secondary_peer_address_prefix" {
+  description = "Secondary peer address prefix for Express Route Circuit private peering"
+  type        = string
+  default     = null
+}
+
+variable "express_route_circuit_private_peering_shared_key" {
+  description = "Shared secret key for Express Route Circuit Private Peering"
+  type        = string
+  default     = null
+}
+
+variable "express_route_circuit_private_peering_vlan_id" {
+  description = "VLAN Id for Express Route Circuit"
+  type        = number
+  default     = null
+}
+
+variable "express_route_circuit_private_peering_peer_asn" {
+  description = "Peer BGP ASN for Express Route Circuit Private Peering"
+  type        = number
+  default     = null
+}
+
+variable "virtual_hub_id" {
+  description = "Id of the Virtual Hub in which to deploy the Firewall"
+  type        = string
+}