+++ modules/virtual-hub/output.tf
@@ -1,0 +1,4 @@
+output "virtual_hub_id" {
+  description = "Id of the virtual hub"
+  value       = azurerm_virtual_hub.vhub.id
+}