+++ r-express-route.tf
@@ -1,40 +1,34 @@
-resource "azurerm_express_route_gateway" "ergw" {
-  for_each            = var.express_route_enabled ? toset(["express_route"]) : toset([])
-  name                = local.ergw_name
-  resource_group_name = var.resource_group_name
-  location            = var.location
-  scale_units         = var.express_route_gateway_scale_unit
-  virtual_hub_id      = azurerm_virtual_hub.vhub.id
+module "express_route" {
+  for_each = var.express_route_enabled ? toset(["express_route"]) : toset([])
+  source   = "./modules/express-route"
 
-  tags = merge(local.tags, var.express_route_gateway_exta_tags)
-}
+  client_name = var.client_name
+  environment = var.environment
+  stack       = var.stack
 
-resource "azurerm_express_route_circuit" "erc" {
-  for_each            = var.express_route_enabled ? toset(["express_route"]) : toset([])
-  name                = local.erc_name
   location            = var.location
+  location_short      = var.location_short
   resource_group_name = var.resource_group_name
 
-  bandwidth_in_mbps     = var.express_route_circuit_bandwidth_in_mbps
-  peering_location      = var.express_route_circuit_peering_location
-  service_provider_name = var.express_route_circuit_service_provider
+  logs_destinations_ids = var.logs_destinations_ids
 
-  sku {
-    tier   = var.express_route_sku.tier
-    family = var.express_route_sku.family
-  }
+  virtual_hub_id = module.vhub.virtual_hub_id
 
-  tags = merge(local.tags, var.express_route_gateway_exta_tags)
-}
+  express_route_gateway_scale_unit  = var.express_route_gateway_scale_unit
+  custom_express_route_gateway_name = var.custom_express_route_gateway_name
 
-resource "azurerm_express_route_circuit_peering" "ercprivatepeer" {
-  for_each                      = var.express_route_enabled && var.express_route_private_peering_enabled ? toset(["express_route"]) : toset([])
-  resource_group_name           = var.resource_group_name
-  express_route_circuit_name    = azurerm_express_route_circuit.erc["express_route"].name
-  peering_type                  = "AzurePrivatePeering"
-  primary_peer_address_prefix   = var.express_route_circuit_private_peering_primary_peer_address_prefix
-  secondary_peer_address_prefix = var.express_route_circuit_private_peering_secondary_peer_address_prefix
-  vlan_id                       = var.express_route_circuit_private_peering_vlan_id
-  shared_key                    = var.express_route_circuit_private_peering_shared_key
-  peer_asn                      = var.express_route_circuit_private_peering_peer_asn
+  express_route_circuit_bandwidth_in_mbps = var.express_route_circuit_bandwidth_in_mbps
+  express_route_circuit_peering_location  = var.express_route_circuit_peering_location
+  express_route_circuit_service_provider  = var.express_route_circuit_service_provider
+  express_route_sku                       = var.express_route_sku
+
+  custom_express_route_circuit_name                                   = var.custom_express_route_circuit_name
+  express_route_private_peering_enabled                               = var.express_route_private_peering_enabled
+  express_route_circuit_private_peering_primary_peer_address_prefix   = var.express_route_circuit_private_peering_primary_peer_address_prefix
+  express_route_circuit_private_peering_secondary_peer_address_prefix = var.express_route_circuit_private_peering_secondary_peer_address_prefix
+  express_route_circuit_private_peering_vlan_id                       = var.express_route_circuit_private_peering_vlan_id
+  express_route_circuit_private_peering_shared_key                    = var.express_route_circuit_private_peering_shared_key
+  express_route_circuit_private_peering_peer_asn                      = var.express_route_circuit_private_peering_peer_asn
+
+  express_route_gateway_tags = merge(local.tags, var.express_route_gateway_exta_tags)
 }